refactor: 사용되지 않는 dto 삭제 (#5)
Refs: #5
