--- BACK/spring-app/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.application.name=spring-app
-
-spring.datasource.url=jdbc:h2:tcp://localhost/~/starchive
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-spring.h2.console.enabled=true
