--- BACK/spring-app/src/test/java/com/starchive/springapp/SpringAppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.starchive.springapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringAppApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
